[1.1.11] Optimize UDF function logic (#131)
* Code optimization

* add   udf function createtime && updatetime column

* add udf updatetime sql

+++ linkis-public-enhancements/linkis-basedata-manager/src/main/java/org/apache/linkis/basedatamanager/server/restful/UdfTreeRestfulApi.java
@@ -17,11 +17,17 @@
 
 package org.apache.linkis.basedatamanager.server.restful;
 
+import org.apache.linkis.basedatamanager.server.conf.UdfTreeConf;
+import org.apache.linkis.basedatamanager.server.domain.UdfBaseInfoEntity;
 import org.apache.linkis.basedatamanager.server.domain.UdfTreeEntity;
+import org.apache.linkis.basedatamanager.server.service.UdfBaseInfoService;
 import org.apache.linkis.basedatamanager.server.service.UdfTreeService;
 import org.apache.linkis.common.conf.Configuration;
 import org.apache.linkis.server.Message;
 import org.apache.linkis.server.utils.ModuleUserUtils;
+
+import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -32,8 +38,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.github.pagehelper.PageInfo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
@@ -46,6 +55,8 @@ import io.swagger.annotations.ApiOperation;
 public class UdfTreeRestfulApi {
 
   @Autowired UdfTreeService udfTreeService;
+
+  @Autowired UdfBaseInfoService udfBaseinfoService;
 
   @ApiImplicitParams({
     @ApiImplicitParam(paramType = "query", dataType = "string", name = "searchName"),
@@ -62,12 +73,33 @@ public class UdfTreeRestfulApi {
     return Message.ok("").data("list", pageList);
   }
 
+  @ApiImplicitParams({
+    @ApiImplicitParam(paramType = "query", dataType = "string", name = "searchName"),
+    @ApiImplicitParam(paramType = "query", dataType = "string", name = "category")
+  })
   @ApiOperation(value = "all", notes = "Query all data of UDF Tree", httpMethod = "GET")
   @RequestMapping(path = "/all", method = RequestMethod.GET)
-  public Message all(HttpServletRequest request, String searchName) {
+  public Message all(HttpServletRequest request, String searchName, String category) {
     ModuleUserUtils.getOperationUser(
         request, "Query all data of UDF Tree,search name:" + searchName);
-    List<UdfTreeEntity> udfTreeEntityList = udfTreeService.list();
+    List<UdfTreeEntity> udfTreeEntityList = new ArrayList<>();
+    UdfTreeEntity entity = new UdfTreeEntity();
+    entity.setCategory(category);
+    QueryWrapper<UdfTreeEntity> querySysWrapper =
+        new QueryWrapper<>(entity)
+            .eq("category", entity.getCategory())
+            .in(
+                "user_name",
+                Arrays.asList(UdfTreeConf.UDF_FUN_SYSTEM_CATEGORY.getValue().split(",")));
+    udfTreeEntityList = udfTreeService.list(querySysWrapper);
+    if (StringUtils.isNotBlank(searchName) && StringUtils.isNotBlank(category)) {
+      entity.setUserName(searchName);
+      QueryWrapper<UdfTreeEntity> queryWrapper =
+          new QueryWrapper<>(entity)
+              .eq("user_name", entity.getUserName())
+              .eq("category", entity.getCategory());
+      udfTreeEntityList.addAll(udfTreeService.list(queryWrapper));
+    }
     return Message.ok("").data("list", udfTreeEntityList);
   }
 
@@ -100,9 +132,27 @@ public class UdfTreeRestfulApi {
   @ApiOperation(value = "remove", notes = "Remove a UDF Tree Record by id", httpMethod = "DELETE")
   @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
   public Message remove(HttpServletRequest request, @PathVariable("id") Long id) {
-    ModuleUserUtils.getOperationUser(request, "Remove a UDF Tree Record,id:" + id.toString());
-    boolean result = udfTreeService.removeById(id);
-    return Message.ok("").data("result", result);
+    String username =
+        ModuleUserUtils.getOperationUser(request, "Remove a UDF Tree Record,id:" + id.toString());
+    if (!Configuration.isAdmin(username)) {
+      return Message.error("User '" + username + "' is not admin user[非管理员用户]");
+    }
+    UdfTreeEntity entity = udfTreeService.getById(id);
+    if (null != entity && entity.getParent() == -1) {
+      return Message.error("The root directory is forbidden to delete[\"根目录禁止删除\"]");
+    }
+    QueryWrapper<UdfTreeEntity> queryWrapper =
+        new QueryWrapper<>(new UdfTreeEntity()).eq("parent", id);
+    List<UdfTreeEntity> folderList = udfTreeService.list(queryWrapper);
+    QueryWrapper<UdfBaseInfoEntity> udfQueryWrapper =
+        new QueryWrapper<>(new UdfBaseInfoEntity()).eq("tree_id", id);
+    List<UdfBaseInfoEntity> functoinList = udfBaseinfoService.list(udfQueryWrapper);
+    if (CollectionUtils.isEmpty(folderList) && CollectionUtils.isEmpty(functoinList)) {
+      boolean result = udfTreeService.removeById(id);
+      return Message.ok("").data("result", result);
+    } else {
+      return Message.error("Please delete the subdirectory first[请先删除子目录]");
+    }
   }
 
   @ApiImplicitParams({
